[Update] Move request folder for clean architecture
