--- BandCommunity.Domain/DTO/Auth/ForgotPasswordRequest.cs
@@ -1,6 +1,0 @@
-﻿namespace BandCommunity.Domain.DTO.Auth;
-
-public class ForgotPasswordRequest
-{
-    public string Email { get; set; }
-}
