--- BandCommunity.Domain/DTO/Auth/LoginRequest.cs
@@ -1,7 +1,0 @@
-﻿namespace BandCommunity.Domain.DTO.Auth;
-
-public class LoginRequest
-{
-    public string Email { get; set; }
-    public string Password { get; set; }
-}
