--- BandCommunity.Domain/DTO/Auth/RegisterRequest.cs
@@ -1,16 +1,0 @@
-﻿namespace BandCommunity.Domain.DTO.Auth;
-
-public class RegisterRequest
-{
-    public string FirstName { get; set; }
-    public string LastName { get; set; }
-    public string Username { get; set; }
-    public string Email { get; set; }
-    public string PhoneNumber { get; set; }
-    public string Address { get; set; }
-    public string Password { get; set; }
-    public string ConfirmPassword { get; set; }
-    public DateTime DateOfBirth { get; set; }
-    public string? ProfilePictureUrl { get; set; }
-    public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
-}
