--- BandCommunity.Domain/DTO/Auth/ResetPasswordRequest.cs
@@ -1,8 +1,0 @@
-﻿namespace BandCommunity.Domain.DTO.Auth;
-
-public class ResetPasswordRequest
-{
-    public string Token { get; set; }
-    public string NewPassword { get; set; }
-    public string ConfirmPassword { get; set; }
-}
