--- BandCommunity.Domain/DTO/User/BasicInfoRequest.cs
@@ -1,6 +1,0 @@
-﻿namespace BandCommunity.Domain.DTO.User;
-
-public class BasicInfoRequest
-{
-    public string Email { get; set; }
-}
