+++ BandCommunity.Domain/Models/Auth/ForgotPasswordRequest.cs
@@ -1,0 +1,9 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace BandCommunity.Domain.Models.Auth;
+
+public class ForgotPasswordRequest
+{
+    [Required]
+    public string Email { get; set; } = null!;
+}
