+++ BandCommunity.Domain/Models/Auth/RegisterRequest.cs
@@ -1,0 +1,26 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace BandCommunity.Domain.Models.Auth;
+
+public class RegisterRequest
+{
+    [Required] 
+    public string FirstName { get; set; } = null!;
+    [Required]
+    public string LastName { get; set; } = null!;
+    [Required]
+    public string Username { get; set; } = null!;
+    [Required]
+    public string Email { get; set; } = null!;
+    [Required]
+    public string PhoneNumber { get; set; } = null!;
+    [Required]
+    public string Address { get; set; } = null!;
+    [Required]
+    public string Password { get; set; } = null!;
+    [Required]
+    public string ConfirmPassword { get; set; } = null!;
+    public DateTime DateOfBirth { get; set; }
+    public string? ProfilePictureUrl { get; set; }
+    public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
+}
