+++ BandCommunity.Domain/Models/User/BasicInfoRequest.cs
@@ -1,0 +1,6 @@
+﻿namespace BandCommunity.Domain.Models.User;
+
+public class BasicInfoRequest
+{
+    public string? Email { get; set; }
+}
